factor out DiffRow/SkipRow
